Rename EventListener to Event directory
