+++ src/Framework/Bootstrap/GacelaConfig.php
@@ -10,7 +10,7 @@ use Gacela\Framework\Config\ConfigReaderInterface;
 use Gacela\Framework\Config\GacelaConfigBuilder\ConfigBuilder;
 use Gacela\Framework\Config\GacelaConfigBuilder\MappingInterfacesBuilder;
 use Gacela\Framework\Config\GacelaConfigBuilder\SuffixTypesBuilder;
-use Gacela\Framework\EventListener\GacelaEventInterface;
+use Gacela\Framework\Event\GacelaEventInterface;
 
 final class GacelaConfig
 {
