+++ src/Framework/Bootstrap/SetupGacelaInterface.php
@@ -7,7 +7,7 @@ namespace Gacela\Framework\Bootstrap;
 use Gacela\Framework\Config\GacelaConfigBuilder\ConfigBuilder;
 use Gacela\Framework\Config\GacelaConfigBuilder\MappingInterfacesBuilder;
 use Gacela\Framework\Config\GacelaConfigBuilder\SuffixTypesBuilder;
-use Gacela\Framework\EventListener\EventDispatcherInterface;
+use Gacela\Framework\Event\EventDispatcherInterface;
 
 interface SetupGacelaInterface
 {
