+++ src/Framework/Event/ClassResolver/ResolvedClassCachedEvent.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Gacela\Framework\EventListener\ClassResolver;
+namespace Gacela\Framework\Event\ClassResolver;
 
 final class ResolvedClassCachedEvent extends AbstractGacelaClassResolverEvent
 {
