+++ src/Framework/Event/ClassResolver/ResolvedClassCreatedEvent.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Gacela\Framework\EventListener\ClassResolver;
+namespace Gacela\Framework\Event\ClassResolver;
 
 final class ResolvedClassCreatedEvent extends AbstractGacelaClassResolverEvent
 {
