+++ src/Framework/Event/GacelaEventInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Gacela\Framework\EventListener;
+namespace Gacela\Framework\Event;
 
 interface GacelaEventInterface
 {
