+++ src/Framework/Event/NullEventDispatcher.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Gacela\Framework\EventListener;
+namespace Gacela\Framework\Event;
 
 final class NullEventDispatcher implements EventDispatcherInterface
 {
